pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-routines"
version = "1.6.0"
description = "Define named groups of management commands in Django settings files for batched execution."
requires-python = ">=3.9,<4.0"
authors = [
  {name = "Brian Kohan", email = "bckohan@gmail.com"},
]
license = "MIT"
license-files = [ "LICENSE" ]
readme = "README.md"
repository = "https://github.com/bckohan/django-routines"
homepage = "https://django-routines.readthedocs.io"
keywords = ["django", "CLI", "management", "routine", "routines", "batch", "commands", "deploy", "tasks", "package", "backup"]
dependencies = [
    "Django>=3.2,<6.0",
    "django-typer>=3.2.2,<4.0"
]
classifiers = [
    "Environment :: Console",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/django_routines"]

[project.optional-dependencies]
# this should track typer's rich dependency, so long as our console
# patches still work - so be sure to test on the low end of the range
rich = ["rich>=10.11.0,<14.0.0"]

[project.urls]
"Homepage" = "https://django-routines.readthedocs.io"
"Documentation" = "https://django-routines.readthedocs.io"
"Repository" = "https://github.com/bckohan/django-routines"
"Issues" = "https://github.com/bckohan/django-routines/issues"
"Changelog" = "https://django-routines.readthedocs.io/en/latest/changelog.html"
"Code_of_Conduct" = "https://github.com/bckohan/django-routines/blob/main/CODE_OF_CONDUCT.md"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest-django>=4.7.0",
    "pytest-cov>=5.0.0",
    "aiohttp>=3.9.1",
    "pytest-env>=1.0.0",
    "pexpect>=4.9.0",
    "pyright>=1.1.357",
    "ruff>=0.9",
    "django-stubs>=5.0.2",
    "mypy>=1.10.0",
    "pre-commit>=4.0.1",
    "django-render-static>=3.2.1",
    "pywinpty>=2.0.14; sys_platform == 'win32'",
    "tomlkit>=0.13.2",
    "readme-renderer[md]>=44.0",
    "packaging>=24.2",
]
docs = [
    "doc8>=1.1.2",
    "furo>=2024.8.6",
    "sphinx>=7.4.7",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-tabs>=3.4.7",
    "sphinxcontrib-django>=2.5",
    "sphinxcontrib-typer[png]>=0.5.1",
]

[tool.mypy]
# The mypy configurations: http://bit.ly/2zEl9WI
# todo some stubs errors
ignore_missing_imports = true
exclude = ["tests"]
mypy_path = "src"
modules = "django_routines"


# todo doc8 configuration here is not being picked up and doesnt seem to be working
# as expected - for now pass these parameters manually
[tool.doc8]
max-line-length = 100
sphinx = true

[tool.pytest.ini_options]
# py.test options:
DJANGO_SETTINGS_MODULE = "tests.settings"
PYTHONIOENCODING = "utf-8"
python_files = "test_*.py"
norecursedirs = "*.egg .eggs dist build docs .tox .git __pycache__ shellcompletion"
env = [
    "TERMINAL_WIDTH=80",
]

addopts = [
  "--strict-markers",
  "--cov=django_routines",
  "--cov-branch",
  "--cov-report=term-missing:skip-covered"
]

[tool.coverage.run]
omit = ["tests/**/*.py"]
branch = true
source = ["django_routines"]
concurrency = ["multiprocessing"]
parallel = true
relative_files = true
command_line = "-m pytest --cov=django_routines"

[tool.coverage.paths]
source = [
    "src/django_routines"
]

[tool.pyright]
exclude = ["tests/**/*"]
include = [
    "src/django_routines"
]

[tool.ruff]
line-length = 88
exclude = [
    "doc",
    "dist",
    "examples"
]

[tool.ruff.lint]
exclude = [
    "tests/**/*"
]
